[BACKLOG-40217] updated vfs-hdfs and added back blueprint for kettle-pluging hdfs step
