+++ impl/vfs/hdfs/assemblies/plugin/pom.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <artifactId>vfs-hdfs-assemblies</artifactId>
+    <groupId>pentaho</groupId>
+    <version>10.2.0.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>pdi-vfs-hdfs-plugin</artifactId>
+  <packaging>pom</packaging>
+
+  <name>PDI Vfs Hdfs Plugin Distribution</name>
+
+  <properties>
+    <resources.directory>${project.basedir}/src/main/resources</resources.directory>
+    <assembly.dir>${project.build.directory}/assembly</assembly.dir>
+  </properties>
+
+  <dependencies>
+    <dependency>
+      <groupId>pentaho</groupId>
+      <artifactId>pentaho-big-data-impl-vfs-hdfs</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+  </dependencies>
+</project>
