+++ impl/vfs/hdfs/assemblies/plugin/src/assembly/assembly.xml
@@ -1,0 +1,51 @@
+<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.3"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.3 http://maven.apache.org/xsd/assembly-1.1.3.xsd">
+  <id>zip</id>
+  <formats>
+    <format>zip</format>
+  </formats>
+
+  <baseDirectory>pdi-vfs-hfds-plugin</baseDirectory>
+
+  <fileSets>
+    <fileSet>
+      <directory>${resources.directory}</directory>
+      <outputDirectory>.</outputDirectory>
+      <filtered>true</filtered>
+    </fileSet>
+
+    <!-- the staging dir -->
+    <fileSet>
+      <directory>${assembly.dir}</directory>
+      <outputDirectory>.</outputDirectory>
+    </fileSet>
+  </fileSets>
+
+  <dependencySets>
+    <dependencySet>
+      <outputDirectory>.</outputDirectory>
+      <includes>
+        <include>pentaho:pentaho-big-data-impl-vfs-hdfs:jar</include>
+      </includes>
+      <useProjectArtifact>false</useProjectArtifact>
+    </dependencySet>
+    <dependencySet>
+      <scope>runtime</scope>
+      <outputDirectory>.</outputDirectory>
+      <useTransitiveDependencies>false</useTransitiveDependencies>
+      <useProjectArtifact>false</useProjectArtifact>
+      <includes>
+        <include>pentaho:pentaho-big-data-impl-vfs-hdfs:jar</include>
+      </includes>
+    </dependencySet>
+    <dependencySet>
+      <scope>runtime</scope>
+      <useProjectArtifact>false</useProjectArtifact>
+      <outputDirectory>lib</outputDirectory>
+      <excludes>
+        <exclude>pentaho:pentaho-big-data-impl-vfs-hdfs:*</exclude>
+      </excludes>
+    </dependencySet>
+  </dependencySets>
+</assembly>
