+++ impl/vfs/hdfs/assemblies/plugin/src/main/resources/version.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<version branch='TRUNK'>${project.version}</version>
