+++ impl/vfs/hdfs/assemblies/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>pentaho</groupId>
+    <artifactId>pentaho-big-data-impl-vfs-hdfs</artifactId>
+    <version>10.2.0.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>vfs-hdfs-assemblies</artifactId>
+  <packaging>pom</packaging>
+
+  <name>PDI Vfs Hdfs Plugin Assemblies</name>
+
+  <modules>
+    <module>plugin</module>
+  </modules>
+</project>
