+++ impl/vfs/hdfs/core/src/main/java/org/pentaho/big/data/impl/vfs/hdfs/nc/NamedClusterProvider.java
