+++ impl/vfs/hdfs/core/src/main/resources/OSGI-INF/blueprint/blueprint.xml
