+++ impl/vfs/hdfs/core/src/test/java/org/pentaho/big/data/impl/vfs/hdfs/HDFSFileObjectTest.java
