+++ impl/vfs/pom.xml
@@ -1,15 +1,17 @@
 <?xml version="1.0"?>
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>pentaho</groupId>
     <artifactId>pentaho-big-data-impl</artifactId>
     <version>10.2.0.0-SNAPSHOT</version>
   </parent>
+
   <artifactId>pentaho-big-data-impl-vfs</artifactId>
   <version>10.2.0.0-SNAPSHOT</version>
   <packaging>pom</packaging>
+
   <modules>
     <module>hdfs</module>
   </modules>
